Not use use_system_binary for system xmlrpc-epi
The xmlrpc-epi package has no .pc or .cmake files. On some platforms,
the header and the library directories don't have special paths. On
GNU/Linux, at least on Debian, the headers are encapsulated in the
packages's own directory. On macOS, both MacPorts & Homebrew don't have
the package. On the other hand, Fink, that has the package, still
doesn't support recent versions of macOS as of this writing. So it's
very likely that on macOS, xmlrpc-epi is installed in /usr/local.