LL physics extensions stub
Since the CMakeLists.txt includes some same .cmake files as the viewer,
I think the project might as well be a part of the Linden libraries
code. And for now is put under llprimitive (might not be consistent, in
fact the opposite, with they way llplugin relates to slplugin), but I
think this way results the least change, and it still works.

The differences include:
- all files (common llphysicsextensions headers to be included by
  library users and the stub implementation files) are put inside one
  directory, and the CMakeLists.txt is adjusted accordingly;
- modernised CMakeLists.txt, so include_directories are now implied by
  target_link_libraries;
- some file name fix;
- add_library is not explicitly set to STATIC;