Deprecated declarations aren't treated as errors
I had added this to CMAKE_CXX_FLAGS in 00-Common before, and only when
the compiler was Clang. But it turned out that GCC was treating them as
errors too, that the addition would need to be applied to all compilers.
So I prefer to put it here in llcommon with the scope set to PUBLIC
cause the errors would show up again when compiling other LL libraries
if the scope is set to something else.