Undefine NULL before defining it to 0 on FreeBSD
Since C++11, NULL is promoted to nullptr on some BSD platforms. This is
very problematic when used with Boost. At times it would fail during
compile-time. What's worse is if it passes compile-time, but then crash
during run-time, for example when some condition is to be checked for its
truth, when then it would be compared to a nullptr.