Fix what GCC considers as misleading indentations
The style conventions aren't really being followed that the different
styles of using tabs or spaces as indentations lead to GCC considering
them as misleading.
It's better to just fix them (but as little as possible as to minimise
this fork difference from upstream) than to supress the warnings from
being treated as errors.