Pass texture entry pointer instead to copyTexture
Otherwise on GCC the -Waddress & -Wno-nonnull-compare warnings would be
treated as errors. It's a different term on Clang, which is undefined
bool conversion. On Clang it's just a warning, but on AppleClang it's
treated as an error.
Rather than suppressing these warnings, I think it's better to just
adjust the interface of copyTexture, since it isn't called that many
times.