Fix FreeBSD login problem by checking endianness
Even though the account was logged in, it would get stuck at getting
region handshake. The problem was because the viewer wasn't getting the
acknowledgement to the successfully sent UseCircuitCode message. I
compared the message data, and it differed (from Linux) on the byte
order of the Code variable (the SessionID & agent ID were right). The
bytes sent to the network weren't reversed (and I was on an Intel
processor).