libmedia_plugin_libvlc is installed to system too
libmedia_plugin_base is now explicitly set as static, since it could be
set as shared when BUILD_SHARED_LIBS is on and it's just easier to have
it compiled into with every media plugin cause it's small and would only
be shared by 2 dynamic libraries anyway. Trying to put a dynamic version
where the media plugin can find it so the reference is not broken, seems
like a hassle.