Add sysroot to GLH & Meshoptimizer include paths
When cross-compiling, the host's /usr/local/include would be unsafely
included before. The problem with this was that it leaked other host
library headers unexpectedly, like Boost. The target compilation caught
some function from the host headers which of a newer version, and then
when trying to link to the target libraries, the function wasn't
available yet in the older version.