Make using the system browser for links works
First, in order for launch_url.sh to be executable, it needs to be
installed as a program.
Secondly, the spawn browser command path needs to be adjusted
accordingly.
And last, add chrome (applies to chromium too on FBSD), to the list of
browser commands to try (so chrome wasn't there :/, but dillo has always
been XD, and that's why it kept opening Dillo here haha).