Fix macOS crash by explicitly ask SDL2 for GL 4.1
The reason it kept crashing was because it is now a requirement for the
OpenGL version to be "detected" as equal or higher than 4.09.
All this time (or maybe since macOS switching to SDL2), Megapahit has been
given only OpenGL 2.1 by macOS (4.1 on SL viewer & Alchemy, though funny
it's 2.1 too somehow on Kokua), and now is fixed by setting the profile
to core 4.1 from SDL2.
The commit is on this branch only cause somehow this fix didn't work
on main branch (it crashed). So 6.6.x is stuck with 2.1 for macOS.
We're moving forward with this branch anyway, which will be merged to main
in the (hopefully near) future.