Revert to using Cocoa instead of SDL2 for macOS
Our SDL2-based code is kept Darwin-ready. This move was triggered
by the merge with SLv's release/maint-b branch causing the viewer
to show only a black screen, even though everything else seemed to
be working (logging in, streaming, quitting the app). This decision
also has caused numerous MPv specific window-related macOS bugs to
be fixed. I wanted to commit this on the main branch, but somehow
it failed to build. This move was bound to happen at the arrival of
Maintenance B after all.