Revert the key type from U32 to U16
It was changed in the first place to accommodate SDL2's bigger size
need, since SDL2 was still used by all ports. This would conflict
when maint-b gets merge later, so just revert it beforehand. With
this commit, it would fail to compile for Linux & FreeBSD on main
branch temporarily.