Wine-based Vivox-based voice support for FreeBSD
https://megapahit.com/show_bug.cgi?id=57
For SLVoice to run, there is a couple of environment variables that need
to be set, and users need to run
`/usr/local/share/wine/pkg32.sh install wine mesa-dri`
to install the necessary files first.
All this time the env parameter on apr_proc_create was never used, and
this is just the place for passing those environment settings, hence the
LLProcess::Param attribute addition.