cURL build instructions that work on AppleClang 15
It has a newer default of something higher than C99, which would
cause an error of implicit int at configuring stage, so it wasn't
about some missing library for runtime.
It's also not necessary to specify the path to nghttp2 (or zlib)
since it seems to have already been detected.
We might as well lower the minimum macOS requirement here to 11.