Fix libllwebrtc.dylib installation to the package
llwebrtc.h is not unnecessarily installed any more.
When code signing the app on macOS, there was also a problem where
it said that this header wasn't signed. ???
That's why we make sure that it won't be installed now.
Also on macOS, the dylib is installed to Frameworks right away, so
now we don't have any unused double in Resources.