Build process' set up to link to Boost statically
on macOS and at least the one directly. Collada DOM's Boost dependency
is still 1.76 in MacPorts' case, and that's why we still have Boost
filesystem and system dylibs in Frameworks. On the other hand, the
viewer codebase now really depends on newer Boost, in my case I can
use MacPorts' 1.81. I had to switch to static because Boost 1.81
filesystem crashed for not finding the implementation of something
declared using BOOST_FORCEINLINE in boost/filesystem/path.hpp.
I think I know why, now. Cause the filesystem dylib that eventually
got installed was the 1.76 one depended on by Collada DOM, so there
was a conflict, there. For now the temporary MacPorts solution for
this is to install boost181 with -no_static variant (notice the "-"
there, so the static libraries are built and installed too). The
rest is so hack-ish, I had to manually recreate Boost links pointing
to 1.81 ones, only the ones needed, and for the libraries, only the
static ones.