Automatic XMLRPC-EPI Mac download, build & install
When linking to the static archive on arm64, it would fail with an
error of undefined iconv symbols.
When linking to the dynamic library, the path was to the dynamic
library was still assumed to still be in /usr/local/lib, so specifically
for the executable target, this needs to be changed (temporarily) to
the one in the prebuilt directory first, to later be changed again
by fixup_bundle to the bundled one in Frameworks.