Stream notification
https://megapahit.com/show_bug.cgi?id=56
I could make such metadata setting triggered by an event only on the
media plugin's side, not on the viewer's side. I had tried adding
a MEDIA_EVENT_NOWPLAYING_CHANGED and hoped that mediactrl, viewerparcelmedia,
or viewermedia would react to it, but no, so this is the best I could
come up with for now, to keep checking if what's now playing has been
updated or not. I intentionally didn't clear mNowPlaying the way mURL is,
so that it wouldn't notify one last double after the player stops streaming.
mNowPlaying and mMediaNowPlaying need to be kept in their last states so
that the comparison is correct. I also intentionally didn't put the
notification inside the MEDIA_PLAYING scope, cause the flow somehow never
got into that scope.