Automatic codesigning on macOS
CMAKE_OSX_DEPLOYMENT_TARGET here, even though reset in Variables.cmake
with mmacosx-version-min, will be used as the hardened runtime version
when codesigning.
Instructions use 11 as that version, as the builder is assumed to be
building for arm64. When building for x86-64, you can replace all 11
here with 10.15.
The sudo in codesigning is required for builders on Apple Silicon whose
SIP is enabled, which is assumed to be the most likely case.
Credits to Cate (32a).